quickply/requirements.txt
Babel==2.9.0
beautifulsoup4==4.9.3
blinker==1.4
bs4==0.0.1
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
DateTime==4.3
dnspython==2.1.0
email-validator==1.1.2
filelock==3.0.12
flash==1.0.3
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.29.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
greenlet==1.0.0
httplib2==0.19.1
idna==2.10
importlib-metadata==4.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
lxml==4.6.3
MarkupSafe==1.1.1
mongoengine==0.23.0
mysql-connector-python==8.0.23
numpy==1.20.2
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.9
pandas==1.2.4
passlib==1.7.4
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.11.3
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.45
scikit-learn==0.24.1
scipy==1.6.3
sentencepiece==0.1.91
six==1.15.0
sklearn==0.0
soupsieve==2.2.1
speaklater==1.3
SQLAlchemy==1.4.9
threadpoolctl==2.1.0
tokenizers==0.9.3
torch==1.4.0
tqdm==4.60.0
transformers==3.5.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
waitress==2.0.0
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.1
zope.interface==5.4.0
